........
  r11346 | martineau | 2008-05-21 10:00:45 -0400 (Wed, 21 May 2008) | 6 lines
  
  BugId: 1256
  * server-src/dumper.c server-src/planner.c, server-src/chunker.c,
    server-src/amflush.c, server-src/driver.c, server-src/taper.c,
    restore-src/amidxtaped.c, restore-src/restore.c,
    restore-src/amfetchdump.c: Convert all pid_t to long before printing.
........


git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@1038 a8d146d6-cc15-0410-8900-af154a0219e0

+++ restore-src/amfetchdump.c
@@ -421,7 +421,7 @@ main(
 	char *process_name = get_master_process(rst_conf_logfile);
 	error(_("%s exists: %s is already running, or you must run amcleanup"), rst_conf_logfile, process_name);
     }
-    log_add(L_INFO, "%s pid %d", get_pname(), getpid());
+    log_add(L_INFO, "%s pid %ld", get_pname(), (long)getpid());
     search_tapes(NULL, stdin, rst_flags->alt_tapedev == NULL,
                  needed_tapes, dumpspecs, rst_flags, NULL);
     cleanup();
@@ -442,7 +442,7 @@ cleanup(void)
 {
     if (parent_pid == getpid()) {
 	if (get_lock) {
-	    log_add(L_INFO, "pid-done %d\n", getpid());
+	    log_add(L_INFO, "pid-done %ld\n", (long)getpid());
 	    unlink(rst_conf_logfile);
 	}
     }
